Sends email when appointment is canceled
We make use of a library to send the email:
yarn add nodemailer

We create a file at /src/config/mail.js to store some configurations info.

Then we make the class /src/lib/Mail.js to serve as a wrapper for the library and to properly load the configuration into the library so we do not need to keep loading it everytime we have to use it.