Configure nodemailer to use handlebars
Handlebars is a template engine for JS. To use it we have to install two libraries that integrate it with express and nodemailer:

yarn add express-handlebars nodemailer-express-handlebars

Then we have to change our Mail class to configure it. Handlebars needs to know where we will be saving our template files (layouts, partials and templates). With that configured it, we just tell nodemailer to render emails using handlebars.