Creates queue to deliver emails
We used redis to manage a queue of emails:

docker run --name redisbarber -p 6379:6379 -d -t redis:alpine

And a library called bee-queue to manage the queue itself making use of redis:

yar add bee-queue

We had to create a class called Queue.js to wrap the bee-queue library. It is used
to deliver new jobs to queues, where each queue is a different kind of email that can be sent.