Carryng on gobarber api
* Creates migration and model for appointments

Do not forget about this commands while working with sequelize:
yarn sequelize migrate:create --name=name-of-migration
yarn sequelize db:migrate

* Creates AppointmentController to store appointments

* Validates date passed to create an appointment

Used the date-fns library
yarn add date-fns@next

The instructor recomended getting the @next tag of the library

* Lists appointments for the current logged user

* Adds pagination to appointments list

* Creates /scchedules route to list appointmentes for providers

* Configures mongo db in the application

We are using docker to user mongo db

docker run --name mongobarber -p 27017:27017 -d -t mongo

And mongoose to access it

yar add mongoose

* Notifyes provider about a new appointment

* Lists notifications for a provider

* [Extra] Do not allow provider to create appointment with himself

* Create route to mark notificaitons as read

* Creates route to delete appointments

* Sends email when appointment is canceled

We make use of a library to send the email:
yarn add nodemailer

We create a file at /src/config/mail.js to store some configurations info.

Then we make the class /src/lib/Mail.js to serve as a wrapper for the library and to properly load the configuration into the library so we do not need to keep loading it everytime we have to use it.

* Configure nodemailer to use handlebars

Handlebars is a template engine for JS. To use it we have to install two libraries that integrate it with express and nodemailer:

yarn add express-handlebars nodemailer-express-handlebars

Then we have to change our Mail class to configure it. Handlebars needs to know where we will be saving our template files (layouts, partials and templates). With that configured it, we just tell nodemailer to render emails using handlebars.

* Creates queue to deliver emails

We used redis to manage a queue of emails:

docker run --name redisbarber -p 6379:6379 -d -t redis:alpine

And a library called bee-queue to manage the queue itself making use of redis:

yar add bee-queue

We had to create a class called Queue.js to wrap the bee-queue library. It is used
to deliver new jobs to queues, where each queue is a different kind of email that can be sent.

* Logs errors on queue

* Creates route to get available times

* Adds more information to Appointments

* Integrates Sentry and treats exceptions

Sentry is a nice servie to capture and track exceptions in the application

yarn add @sentry/node@5.6.2

Also we have to tell express to properly pass down the exceptions inside a async function

yarn add express-async-errors

Then we use Youch to format any exception that happens while the application runs

yarn add youch

* Adds dotenv to the project